project/backend/service/requirements.txt
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
ffmpeg==1.4
grpcio==1.38.0
grpcio-tools==1.38.0
idna==2.10
imageio==2.9.0
netifaces==0.10.6
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
protobuf==3.17.1
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.25.1
scenedetect==0.5.5
six==1.16.0
socketio==0.2.1
tqdm==4.61.0
urllib3==1.26.5
websocket-client==1.2.1
